# internal/openapi/openapi.json
{
  "components": {
    "responses": {
...
      "CustomerConflict": {
        "content": {
          "application/json": {
            "example": {
              "error": {
                "code": "customer_already_exists",
                "message": "This customer ID already exists in the selected environment."
              }
            },
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        },
        "description": "The customer ID already exists in the selected environment."
      },
...
      "CustomerFailed": {
        "content": {
          "application/json": {
            "example": {
              "error": {
                "code": "customer_operation_failed",
                "message": "Consumel could not complete the customer request. Try again shortly."
              }
            },
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        },
        "description": "The customer request could not be completed."
      },
...
      "CustomerListInvalid": {
        "content": {
          "application/json": {
            "examples": {
              "invalid_cursor": {
                "value": {
                  "error": {
                    "code": "invalid_cursor"
                  }
                }
              },
              "invalid_request": {
                "value": {
                  "error": {
                    "code": "invalid_request"
                  }
                }
              },
              "invalid_search": {
                "value": {
                  "error": {
                    "code": "invalid_search"
                  }
                }
              }
            },
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        },
        "description": "The list pagination or search input is invalid."
      },
...
      "InvalidAPIKey": {
        "content": {
          "application/json": {
            "example": {
              "error": {
                "code": "invalid_api_key",
                "message": "Provide an active project environment API key."
              }
            },
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        },
        "description": "The project API key is missing, malformed, revoked, replaced, or inactive."
      },
...
    },
    "schemas": {
...
      "Customers": {
        "additionalProperties": false,
        "properties": {
          "customers": {
            "items": {
              "$ref": "#/components/schemas/Customer"
            },
            "type": "array"
          },
          "nextCursor": {
            "type": [
              "string",
              "null"
            ]
          }
        },
        "required": [
          "customers",
          "nextCursor"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "localCookieSession": {
        "in": "cookie",
        "name": "consumel_session",
        "type": "apiKey"
      },
      "productionCookieSession": {
        "in": "cookie",
        "name": "__Host-consumel_session",
        "type": "apiKey"
      },
      "projectAPIKey": {
        "bearerFormat": "cm_test_… or cm_live_…",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Consumel usage-based billing infrastructure API.",
    "title": "Consumel API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/customers": {
      "get": {
        "parameters": [
          {
            "description": "A case-insensitive customer ID, name, or email search.",
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "maxLength": 120,
              "type": "string"
            }
          },
          {
            "description": "The opaque next cursor from the previous page.",
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "The number of customers to return.",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 25,
              "maximum": 100,
              "minimum": 1,
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "example": {
                  "customers": [
                    {
                      "createdAt": "2026-09-28T12:00:00Z",
                      "customerId": "user_123",
                      "email": "jordan@example.com",
                      "metadata": {
                        "country": "US",
                        "location": "New York, NY",
                        "plan": "growth"
                      },
                      "name": "Jordan Lee",
                      "updatedAt": "2026-09-28T12:00:00Z"
                    }
                  ],
                  "nextCursor": null
                },
                "schema": {
                  "$ref": "#/components/schemas/Customers"
                }
              }
            },
            "description": "Completed successfully."
          },
          "400": {
            "$ref": "#/components/responses/CustomerListInvalid"
          },
          "401": {
            "$ref": "#/components/responses/InvalidAPIKey"
          },
          "500": {
            "$ref": "#/components/responses/CustomerFailed"
          }
        },
        "security": [
          {
            "projectAPIKey": []
          }
        ],
        "summary": "List customers in the API key's project environment.",
        "tags": [
          "Customers"
        ]
      },
      "post": {
        "requestBody": {
          "content": {
            "application/json": {
              "example": {
                "customerId": "user_123",
                "email": "jordan@example.com",
                "metadata": {
                  "country": "US",
                  "location": "New York, NY",
                  "plan": "growth"
                },
                "name": "Jordan Lee"
              },
              "schema": {
                "$ref": "#/components/schemas/CreateCustomerRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "example": {
                  "createdAt": "2026-09-28T12:00:00Z",
                  "customerId": "user_123",
                  "email": "jordan@example.com",
                  "metadata": {
                    "country": "US",
                    "location": "New York, NY",
                    "plan": "growth"
                  },
                  "name": "Jordan Lee",
                  "updatedAt": "2026-09-28T12:00:00Z"
                },
                "schema": {
                  "$ref": "#/components/schemas/Customer"
                }
              }
            },
            "description": "Created."
          },
          "400": {
            "$ref": "#/components/responses/CustomerInvalid"
          },
          "401": {
            "$ref": "#/components/responses/InvalidAPIKey"
          },
          "409": {
            "$ref": "#/components/responses/CustomerConflict"
          },
          "500": {
            "$ref": "#/components/responses/CustomerFailed"
          }
        },
        "security": [
          {
            "projectAPIKey": []
          }
        ],
        "summary": "Create a customer in the API key's project environment.",
        "tags": [
          "Customers"
        ]
      }
    },
    "/v1/customers/{customerId}": {
      "get": {
        "parameters": [
          {
            "description": "The customer identifier supplied by the integrating application.",
            "in": "path",
            "name": "customerId",
            "required": true,
            "schema": {
              "maxLength": 255,
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "example": {
                  "createdAt": "2026-09-28T12:00:00Z",
                  "customerId": "user_123",
                  "email": "jordan@example.com",
                  "metadata": {
                    "country": "US",
                    "location": "New York, NY",
                    "plan": "growth"
                  },
                  "name": "Jordan Lee",
                  "updatedAt": "2026-09-28T12:00:00Z"
                },
                "schema": {
                  "$ref": "#/components/schemas/Customer"
                }
              }
            },
            "description": "Completed successfully."
          },
          "400": {
            "$ref": "#/components/responses/CustomerInvalid"
          },
          "401": {
            "$ref": "#/components/responses/InvalidAPIKey"
          },
          "404": {
            "$ref": "#/components/responses/CustomerNotFound"
          },
          "500": {
            "$ref": "#/components/responses/CustomerFailed"
          }
        },
        "security": [
          {
            "projectAPIKey": []
          }
        ],
        "summary": "Return one customer from the API key's project environment.",
        "tags": [
          "Customers"
        ]
      },
      "put": {
        "parameters": [
          {
            "description": "The customer identifier supplied by the integrating application.",
            "in": "path",
            "name": "customerId",
            "required": true,
            "schema": {
              "maxLength": 255,
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "example": {
                "email": "jordan@example.com",
                "metadata": {
                  "country": "US",
                  "location": "New York, NY",
                  "plan": "growth"
                },
                "name": "Jordan Lee"
              },
              "schema": {
                "$ref": "#/components/schemas/UpdateCustomerRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "example": {
                  "createdAt": "2026-09-28T12:00:00Z",
                  "customerId": "user_123",
                  "email": "jordan@example.com",
                  "metadata": {
                    "country": "US",
                    "location": "New York, NY",
                    "plan": "growth"
                  },
                  "name": "Jordan Lee",
                  "updatedAt": "2026-09-28T12:00:00Z"
                },
                "schema": {
                  "$ref": "#/components/schemas/Customer"
                }
              }
            },
            "description": "Completed successfully."
          },
          "400": {
            "$ref": "#/components/responses/CustomerInvalid"
          },
          "401": {
            "$ref": "#/components/responses/InvalidAPIKey"
          },
          "404": {
            "$ref": "#/components/responses/CustomerNotFound"
          },
          "500": {
            "$ref": "#/components/responses/CustomerFailed"
          }
        },
        "security": [
          {
            "projectAPIKey": []
          }
        ],
        "summary": "Replace a customer's optional contact information and metadata.",
        "tags": [
          "Customers"
        ]
      }
    },
...
    "/v1/projects/{projectId}/environments/{environment}/customers": {
      "get": {
        "parameters": [
          {
            "description": "The immutable ID of the project selected in the dashboard.",
            "in": "path",
            "name": "projectId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "description": "The selected isolated project environment.",
            "in": "path",
            "name": "environment",
            "required": true,
            "schema": {
              "enum": [
                "sandbox",
                "live"
              ],
              "type": "string"
            }
          },
          {
            "description": "A case-insensitive customer ID, name, or email search.",
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "maxLength": 120,
              "type": "string"
            }
          },
          {
            "description": "The opaque next cursor from the previous page.",
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "The number of customers to return.",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 25,
              "maximum": 100,
              "minimum": 1,
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "example": {
                  "customers": [
                    {
                      "createdAt": "2026-09-28T12:00:00Z",
                      "customerId": "user_123",
                      "email": "jordan@example.com",
                      "metadata": {
                        "country": "US",
                        "location": "New York, NY",
                        "plan": "growth"
                      },
                      "name": "Jordan Lee",
                      "updatedAt": "2026-09-28T12:00:00Z"
                    }
                  ],
                  "nextCursor": null
                },
                "schema": {
                  "$ref": "#/components/schemas/Customers"
                }
              }
            },
            "description": "Completed successfully."
          },
          "400": {
            "$ref": "#/components/responses/CustomerListInvalid"
          },
          "401": {
            "$ref": "#/components/responses/NotAuthenticated"
          },
          "404": {
            "$ref": "#/components/responses/CustomerEnvironmentNotFound"
          },
          "409": {
            "$ref": "#/components/responses/CustomerEnvironmentConflict"
          },
          "500": {
            "$ref": "#/components/responses/CustomerFailed"
          }
        },
        "security": [
          {
            "productionCookieSession": []
          },
          {
            "localCookieSession": []
          }
        ],
        "summary": "List customers in the signed-in project workspace.",
        "tags": [
          "Dashboard Customers"
        ]
      },
      "post": {
        "parameters": [
          {
            "description": "The immutable ID of the project selected in the dashboard.",
            "in": "path",
            "name": "projectId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "description": "The selected isolated project environment.",
            "in": "path",
            "name": "environment",
            "required": true,
            "schema": {
              "enum": [
                "sandbox",
                "live"
              ],
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "example": {
                "customerId": "user_123",
                "email": "jordan@example.com",
                "metadata": {
                  "country": "US",
                  "location": "New York, NY",
                  "plan": "growth"
                },
                "name": "Jordan Lee"
              },
              "schema": {
                "$ref": "#/components/schemas/CreateCustomerRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "example": {
                  "createdAt": "2026-09-28T12:00:00Z",
                  "customerId": "user_123",
                  "email": "jordan@example.com",
                  "metadata": {
                    "country": "US",
                    "location": "New York, NY",
                    "plan": "growth"
                  },
                  "name": "Jordan Lee",
                  "updatedAt": "2026-09-28T12:00:00Z"
                },
                "schema": {
                  "$ref": "#/components/schemas/Customer"
                }
              }
            },
            "description": "Created."
          },
          "400": {
            "$ref": "#/components/responses/CustomerInvalid"
          },
          "401": {
            "$ref": "#/components/responses/NotAuthenticated"
          },
          "404": {
            "$ref": "#/components/responses/CustomerEnvironmentNotFound"
          },
          "409": {
            "$ref": "#/components/responses/CustomerCreateConflict"
          },
          "500": {
            "$ref": "#/components/responses/CustomerFailed"
          }
        },
        "security": [
          {
            "productionCookieSession": []
          },
          {
            "localCookieSession": []
          }
        ],
        "summary": "Create a customer from the signed-in project workspace.",
        "tags": [
          "Dashboard Customers"
        ]
      }
    },
    "/v1/projects/{projectId}/environments/{environment}/customers/{customerId}": {
      "get": {
        "parameters": [
          {
            "description": "The immutable ID of the project selected in the dashboard.",
            "in": "path",
            "name": "projectId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "description": "The selected isolated project environment.",
            "in": "path",
            "name": "environment",
            "required": true,
            "schema": {
              "enum": [
                "sandbox",
                "live"
              ],
              "type": "string"
            }
          },
          {
            "description": "The customer identifier supplied by the integrating application.",
            "in": "path",
            "name": "customerId",
            "required": true,
            "schema": {
              "maxLength": 255,
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "example": {
                  "createdAt": "2026-09-28T12:00:00Z",
                  "customerId": "user_123",
                  "email": "jordan@example.com",
                  "metadata": {
                    "country": "US",
                    "location": "New York, NY",
                    "plan": "growth"
                  },
                  "name": "Jordan Lee",
                  "updatedAt": "2026-09-28T12:00:00Z"
                },
                "schema": {
                  "$ref": "#/components/schemas/Customer"
                }
              }
            },
            "description": "Completed successfully."
          },
          "400": {
            "$ref": "#/components/responses/CustomerInvalid"
          },
          "401": {
            "$ref": "#/components/responses/NotAuthenticated"
          },
          "404": {
            "$ref": "#/components/responses/CustomerOrEnvironmentNotFound"
          },
          "409": {
            "$ref": "#/components/responses/CustomerEnvironmentConflict"
          },
          "500": {
            "$ref": "#/components/responses/CustomerFailed"
          }
        },
        "security": [
          {
            "productionCookieSession": []
          },
          {
            "localCookieSession": []
          }
        ],
        "summary": "Return one customer in the signed-in project workspace.",
        "tags": [
          "Dashboard Customers"
        ]
      },
      "put": {
        "parameters": [
          {
            "description": "The immutable ID of the project selected in the dashboard.",
            "in": "path",
            "name": "projectId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "description": "The selected isolated project environment.",
            "in": "path",
            "name": "environment",
            "required": true,
            "schema": {
              "enum": [
                "sandbox",
                "live"
              ],
              "type": "string"
            }
          },
          {
            "description": "The customer identifier supplied by the integrating application.",
            "in": "path",
            "name": "customerId",
            "required": true,
            "schema": {
              "maxLength": 255,
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "example": {
                "email": "jordan@example.com",
                "metadata": {
                  "country": "US",
                  "location": "New York, NY",
                  "plan": "growth"
                },
                "name": "Jordan Lee"
              },
              "schema": {
                "$ref": "#/components/schemas/UpdateCustomerRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "example": {
                  "createdAt": "2026-09-28T12:00:00Z",
                  "customerId": "user_123",
                  "email": "jordan@example.com",
                  "metadata": {
                    "country": "US",
                    "location": "New York, NY",
                    "plan": "growth"
                  },
                  "name": "Jordan Lee",
                  "updatedAt": "2026-09-28T12:00:00Z"
                },
                "schema": {
                  "$ref": "#/components/schemas/Customer"
                }
              }
            },
            "description": "Completed successfully."
          },
          "400": {
            "$ref": "#/components/responses/CustomerInvalid"
          },
          "401": {
            "$ref": "#/components/responses/NotAuthenticated"
          },
          "404": {
            "$ref": "#/components/responses/CustomerOrEnvironmentNotFound"
          },
          "409": {
            "$ref": "#/components/responses/CustomerEnvironmentConflict"
          },
          "500": {
            "$ref": "#/components/responses/CustomerFailed"
          }
        },
        "security": [
          {
            "productionCookieSession": []
          },
          {
            "localCookieSession": []
          }
        ],
        "summary": "Replace a customer's optional contact information and supported metadata from the signed-in project workspace.",
        "tags": [
          "Dashboard Customers"
        ]
      }
    },
...
  },
  "servers": [
    {
      "description": "Production",
      "url": "https://api.consumel.com"
    },
    {
      "description": "Local development",
      "url": "http://localhost:8080"
    }
  ]
}